+++ format-flowed/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "format-flowed"
+version = "1.0.0"
+description = "format=flowed support"
+edition = "2021"
+license = "MPL-2.0"
+
+keywords = ["email"]
+categories = ["email"]
+
+[dependencies]
